Simplify Example 3 as per issue #82
This change removes all of the HTTP Post functionality from Example 3,
since this appears to cause more confusion than clarification. There are
also some issues with this code:

  https://github.com/w3c/webpayments-payment-apps-api/issues/82#issuecomment-275976677

The remaining code is updated to use promises correctly, to unregister
from the 'message' event correctly, and to make correct use of the
basic-card specification.