Add ImageObject dictionary.
The paymentinstrument and the paymentwallet have a sequence of ImageObject
as member, but there is no definition of ImageObject in the current spec.
We already know implicitly that is is similar to the image object in the
manifest spec[1]. But the does not have WebIDL definition of it, so it is
difficult to represent our data structure. To resolve the issue, this change
is adding ImageObject dictionary in WebIDL level.