[Spec] Remove user-identifiable information from canMakePayment (#404)
* [Spec] Remove user-identifiable information from canMakePayment

The “canmakepayment” service worker event lets the merchant know whether the
user has a card on file in an installed payment app. It silently passes the
merchant's origin and arbitrary data to a service worker from payment app
origin. This cross-origin communication happens on PaymentRequest construction
in JavaScript, does not require a user gesture, and does not show any user
interface. As such, it is a potential source of silent user tracking in a
post-3p cookies world.

* Fix existing bug pointing to wrong modifiers