Remove PaymentInstruments API
PaymentInstruments is the Web API that backs non-JIT install of payment
apps. It was designed with the assumption that the browser would store
the actual payment instrument details, which has not turned out to be
true, and has some privacy leaks. This patch removes the ability to
register a payment handler via paymentManager.instruments API. Only
just-in-time installation is supported from now on.

Chrome tracking bug: https://crbug.com/1327265
Chrome status: https://chromestatus.com/feature/5099285054488576

Overview of changes in the patch:
 * Remove PaymentInstruments interface.
 * Remove no longer used PaymentInstrument and ImageOjbect dictionaries.
 * Mark PaymentManager exposed only on Window (not Worker) to match the
   Chrome implementation.
 * Add a definition for "payer" for symmetry with "payee" definition.
 * Soften the language around the purpose for the "canmakepayment"
   event to match Chrome implementation, where it is being used for
   hasEnrolledInstrument() instead of canMakePayment() and show().
 * Add a promise to the "canmakepayment" example to reduce confusion
   stemmed from many examples that always unconditionally return "true".
 * Raise 2nd level sections to 1st level for clarity.
 * Remove the unused PaymentRequestEvent.fetchedImage internal slot.